+++ src/redux/configurestore.js
@@ -1,5 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
-import postsSlice from "./modules/post.js";
+
+import postsSlice from "./modules/test";
 
 // 리듀서 이름: 리듀서 이름.reducer,
 const store = configureStore({
@@ -7,5 +8,4 @@ const store = configureStore({
     posts: postsSlice,
   },
 });
-
 export default store;
